refactor: refactor ScaledJob to be used with scale_handler (#4707)
