+++ apis/keda/v1alpha1/scaledjob_types.go
@@ -142,3 +142,7 @@ func (s ScaledJob) MinReplicaCount() int64 {
 	}
 	return defaultScaledJobMinReplicaCount
 }
+
+func (s *ScaledJob) GenerateIdentifier() string {
+	return GenerateIdentifier("ScaledJob", s.Namespace, s.Name)
+}
